+++ nPopup/InfoItem.cpp
@@ -34,7 +34,7 @@ InfoItem::~InfoItem()
 int InfoItem::GetDesiredWidth(int maxWidth)
 {
     SIZE s;
-    mWindow->GetDesiredSize(maxWidth, mWindow->GetSize().height, &s);
+    mWindow->GetDesiredSize(maxWidth, (int)mWindow->GetSize().height, &s);
     return s.cx;
 }
